backend/requirements/dev.txt
-r base.txt

# Development tools
black>=24.3.0
flake8>=6.1.0
autopep8>=2.3.2
bandit>=1.7.0
ipython>=8.12.0
django-debug-toolbar>=4.2.0
django-silk>=5.4.3
nplusone>=1.0.0
